+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 
 - Option to disable the drag feature per condition and per API call
+- Option to hide visibility of EditMode overlay per Frame with an icon
+  - can be disabled/enabled by api or default values
+- Updated dropdowns to modern style
 
 ## [1] – 2025-11-30